impletmenting on QS
